feat(trip): add trip repository and dummy screen for testing (no real tests)
Modify the `TripsRepositoryFirestore` implementation to convert a Firestore document to a `Trip`.

Changing `DummyScreen` and adding a `DummyScreenViewModel` to provide a temporary UI for manually testing the trip repository functionalities. This screen includes buttons to add a hardcoded trip to Firestore and to fetch and display a trip by its ID.

Add a new `TripsRepositoryProvider`: object to provide a singleton instance of the `TripsRepository`.