feat(api): add pagination to popular activities endpoint
This commit introduces pagination to the `getMostPopularActivities` function in the `ActivityRepository`.

A new `page` parameter has been added to the function, allowing clients to fetch popular activities in paginated results. The default page is set to 0.

This change is reflected in the `ActivityRepositoryMySwitzerland` implementation, which now includes the `page` parameter in the API request to fetch the corresponding data.