test(orderActivities): add end 2 end test to order activities.
The test mocks the mapbox API and checks the validity of the url.
It then answers with a JSON and checks that the solution given by the TSP is correct.
The locations are using real data.
The json was created with through Mapbox playground.
Add a dependency to the build.gradle.kts to mock the Mapbox server.