feat(activity): add estimated time to Activity model
This commit introduces an `estimatedTime` field to the `Activity` data class to store the duration of an activity in seconds.

The `ActivityRepositoryMySwitzerland` has been updated to populate this new field when fetching activities. The start and end timestamps are now provisionally set to the current time, with the intention of calculating them later based on the activity's position in a trip.