test(user): update preferences in emulator tests
This commit updates the hardcoded user preference values in the `UserRepositoryEmulatorTest.kt` file. The previous string-based preferences like "Hiking" and "Sports" have been replaced with enum-style values such as "COUPLE" and "FOODIE".

Additionally, a minor correction was made to ensure the repository instance in one of the tests uses the Firestore emulator instance for consistency.