feat(friends): display user profile pictures in friends list
This commit updates the friends list to show user profile pictures instead of the first letter of their name.

If a user has a profile picture, it will be displayed. Otherwise, a default profile picture is shown. This is achieved by using `AsyncImage` to load the image from a URL.

Additionally, `FriendsListScreenViewModel.kt` has been renamed to `FriendsViewModel.kt` for better clarity and consistency.