refactor(test): change testing to adapt with new code
The getActivityNumberToPull method has been created and made public to facilitate testing, and related tests in ActivityRepositoryMockTest were updated to reflect these changes. A test for parsing activities with no
eededtime was removed as it is now obsolete because an activity can't be pulled if its time is not valid now.