feat(friends): refactor `FriendElement` to use state and event classes and add `isAddMode`
This commit refactors the `FriendElement` composable to use the state and events pattern by introducing `FriendElementState` and `FriendElementActions` classes. This improves clarity by grouping related properties and callbacks.

Additionally, a new `isAddMode` flag has been added to the `FriendElementState`. When this flag is true, an add icon is displayed instead of the usual arrow, allowing the component to be used for adding new friends.

Tests have been updated to reflect these changes and to verify the new `isAddMode` functionality.